adjust rss lat/lon to double (common format across other classes)

+++ source/net/yacy/cora/document/feed/Hit.java
@@ -83,7 +83,7 @@ public interface Hit {
     
     public long getSize();
 
-    public float getLon();
+    public double getLon();
     
-    public float getLat();
+    public double getLat();
 }
